Separate conponents to defferent files
