+++ src/helpers/schemas/index.ts
@@ -1,0 +1,1 @@
+export { feedbackSchema } from './feedback-schema';
